backend/requirements.txt
black==21.6b0
Django==3.0.14
djangosaml2==1.3.2
djangorestframework==3.12.4
